[DF-12526] Fixed warnings on windows
Former-commit-id: a2240e022e8af7506a3f5eba9eade9470f7381c8

+++ Programs/ResourceEditor/Classes/Qt/CubemapEditor/CubeListItemDelegate.cpp
@@ -149,8 +149,8 @@ void CubeListItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& 
         const ListCacheItem& cacheItem = itemIter->second;
         for (size_t iconIndex = 0; iconIndex < cacheItem.icons.size(); ++iconIndex)
         {
-            painter->drawImage(QPoint(faceStartX + iconIndex * (thumbnailSize.width() + FACE_IMAGE_BORDER), y + FACE_IMAGE_BORDER),
-                               *cacheItem.icons[iconIndex]);
+            painter->drawImage(QPoint(static_cast<int>(faceStartX + iconIndex * (thumbnailSize.width() + FACE_IMAGE_BORDER)), 
+                                    static_cast<int>(y + FACE_IMAGE_BORDER)), *cacheItem.icons[iconIndex]);
         }
 
         if (cacheItem.actualSize.size() > 0)
